Make killing a sub-process at build stop as option
PEBakery Nightly have been killing a sub-process created by ShellExecute when a build stop is requested.
However, this method have several severe problems.

1. Killing a process always have a chance to corrupt system.
   Ex) Think about killing an imagex process in the process of mounting wim image.
2. PEBakery cannot kill grandchild process.
   Ex) When a batch file is run with cmd.exe in ShellExecute, killing
   the cmd.exe process does not effectively kill the running command of
   the batch file. One can observe this by stopping the build while
   wimlib-imagex is run by a batch file, which is busy with extracting an wim image.

So, the default behavior of PEBakery is now changed not to kill
sub-process, and wait for it. But if a project developer/user like the
old 'always-killing' behavior, the newly added setting will resurrect it.