Don't warn implicit EndLocal if halt flags are set
SetLocal command should be paired with EndLocal, but PEBakery allows
implicit EndLocal with a warning log. But using Exit or Halt command
can force PEBakery to skip EndLocal, making unnecessary warning log.

To prevent this issue, do not log the warning message even if the
EndLocal was implicitly processed when the halt flags* are set.
*Halt Flags = User abort, Exit command, Halt command, error abort