Added Resource property to ReadInterface and WriteInterface commands.
Allows read/edit of a button's Picture, Image's Picture, and TextFile's file resource. Together with the Attachment editor's advanced view this allows a script author to embed additional resources into the script and swap them on the fly, or remove them with Nil.

Note: For compatibility with Winbuilder the Image and TextFile controls store their resource value in the TEXT property, so they are also able to be modified by WriteInterface,Text, though this is not intuitive, and does not properly set a null value for the resource when set to an EmptyString/Nil. This PR does not address this issue.