Simplify none/wrong image icon to rectangle
UIRenderer.RenderImage was using PackMaterialIcon to show empty image or
an image does not exist. However, those icons were designed to be
square, and did not fit to non-square rectangle. While it was possible
to apply ScaleTransform, it did not work well on Interface Editor.

Thus, use simple border instead to deal with these hassles.