Calculate drag delta with better precision (#155)
This commit tries to solve interface editor's Y=1px shift bug.
In SingleMove, DragCanvas was calculating drag delta with effectively
large arhithmetic error.

AS-IS: delta = newElementPos - uiCtrlPos
TO-BE: delta = dragNowPos - dragStartPos

In AS-IS, newElementPos was correct but uiCtrlPos did not.
Sometimes, uiCtrl.X/Y is not equal to UIElement.X/Y.
Thus, always calculate drag delta with drag click positions.