SyntaxChecker no longer prints coverage percent
With the introduction of section deep inspection, SyntaxChecker is now
able to track nearly every Code/Interface section by its own.
Coverage percent will be always 100% or more, so remove the coverage
percent from the message.

Also fixes the issue that the deep inspect fails to detect a code if a
Macro contained a number on its name.